requirements.txt
Django>=2.0.0,<3
channels>=2.0.2
channels_redis~=2.0
django-registration-redux
django-extensions
drf-access-policy
django-oauth-toolkit
django-cors-headers
django-filter
djangorestframework
drf-access-policy
psycopg2
pytest-django
pytest-asyncio
django-channels-graphql-ws
django-guardian
python-slugify
s3fs
boto3
django-storages
scipy
graphene-django

pandas-profiling
zarr
dask[complete]==2.11.0
lz4
numcodecs
xarray
seaborn
blosc
dask-image
Markdown

# For port provision
docker

# Testing and Linting
ptvsd
pytest-django 
pytest-asyncio
pylint
pylint-django

colorlog